services/news-fetcher/requirements.txt
aiohttp==3.9.1
aiokafka==0.10.0
feedparser==6.0.10
pydantic==2.4.2
PyYAML==6.0.1
redis==5.0.1
uvloop==0.19.0
psycopg2-binary==2.9.9
